chore: extract script from ui.js into module
